feat: reorganize shared/api slice
