+++ src/shared/api/index.ts
@@ -1,1 +1,2 @@
-export * from './api'
+export * from './rest'
+export * from './ws'
