--- src/shared/api/queries/index.ts
@@ -1,3 +1,0 @@
-export * from './access-query'
-export * from './api-query'
-export * from './base-query'
