+++ src/shared/api/rest/index.ts
@@ -1,0 +1,2 @@
+export * from './slice.ts'
+export * from './type'
