+++ src/shared/api/rest/utils/is-auth-response-data.ts
@@ -1,4 +1,4 @@
-import { AuthQueryResponseData, Token } from '../queries/type'
+import { AuthQueryResponseData, Token } from '../queries/type.ts'
 
 
 export const isAuthResponseData = (res: unknown): res is AuthQueryResponseData =>
