--- src/shared/api/utils/index.ts
@@ -1,5 +1,0 @@
-export * from './format-ingredients-response'
-export * from './is-auth-response-data'
-export * from './is-response-data'
-export * from './is-response-error'
-export * from './prepare-post-order-body'
